Harden browser test protocol and build-backup hygiene
